correct errors

+++ scalar/src/lib.rs
@@ -2,7 +2,7 @@ pub fn sum(a: u8, b: u8) -> u8 {
     return a + b;
 }
 
-pub fn diff(a: i32, b: i32) -> i32 {
+pub fn diff(a: i16, b: i16) -> i16 {
     return a - b;
 }
